+++ Runtime/Framework/Framework.cs
@@ -146,6 +146,51 @@ namespace DT.UniStart {
       this.onDestroy.AddOnceListener(() => eventBus.RemoveListener(key, action));
       return action;
     }
+    /// <summary>
+    /// Watch an event bus for events.
+    /// Remove the listener when the object is destroyed.
+    /// </summary>
+    public UnityAction OnceWatch<K>(IEventListener<K> eventBus, K key, UnityAction action) {
+      eventBus.AddOnceListener(key, action);
+      this.onDestroy.AddOnceListener(() => eventBus.RemoveOnceListener(key, action));
+      return action;
+    }
+    /// <summary>
+    /// Watch an event bus for events.
+    /// Remove the listener when the object is destroyed.
+    /// </summary>
+    public UnityAction<T0> OnceWatch<K, T0>(IEventListener<K> eventBus, K key, UnityAction<T0> action) {
+      eventBus.AddOnceListener(key, action);
+      this.onDestroy.AddOnceListener(() => eventBus.RemoveOnceListener(key, action));
+      return action;
+    }
+    /// <summary>
+    /// Watch an event bus for events.
+    /// Remove the listener when the object is destroyed.
+    /// </summary>
+    public UnityAction<T0, T1> OnceWatch<K, T0, T1>(IEventListener<K> eventBus, K key, UnityAction<T0, T1> action) {
+      eventBus.AddOnceListener(key, action);
+      this.onDestroy.AddOnceListener(() => eventBus.RemoveOnceListener(key, action));
+      return action;
+    }
+    /// <summary>
+    /// Watch an event bus for events.
+    /// Remove the listener when the object is destroyed.
+    /// </summary>
+    public UnityAction<T0, T1, T2> OnceWatch<K, T0, T1, T2>(IEventListener<K> eventBus, K key, UnityAction<T0, T1, T2> action) {
+      eventBus.AddOnceListener(key, action);
+      this.onDestroy.AddOnceListener(() => eventBus.RemoveOnceListener(key, action));
+      return action;
+    }
+    /// <summary>
+    /// Watch an event bus for events.
+    /// Remove the listener when the object is destroyed.
+    /// </summary>
+    public UnityAction<T0, T1, T2, T3> OnceWatch<K, T0, T1, T2, T3>(IEventListener<K> eventBus, K key, UnityAction<T0, T1, T2, T3> action) {
+      eventBus.AddOnceListener(key, action);
+      this.onDestroy.AddOnceListener(() => eventBus.RemoveOnceListener(key, action));
+      return action;
+    }
     #endregion
 
     #region Re-expose Fn methods