+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Feat: add `DebugCommandBus`, fix #11.
 - Feat: add `IEventListener/IEventInvoker` as sub-interface of `IEventBus`.
+- Feat: add `UniStartBehaviour.OnceWatch` for `IEventListener`.
 - Note: optimize output format of `DebugEventBus`.
 
 ## v6.0.1